// src/types/pocketbase-types.ts
export enum Collections {
  CategoriaPersona = "CategoriaPersona",
  Cuestionario = "Cuestionario",
  MediaSuperior = "Media_Superior",
  Preguntas = "Preguntas",
  Superior = "Superior",
  TipoBachiller = "TipoBachiller",
  TipoInstitucion = "TipoInstitucion",
  Usuario = "Usuario",
  Users = "users",
}

// Alias types for improved usability
export type IsoDateString = string;
export type RecordIdString = string;
export type HTMLString = string;

// System fields
export type BaseSystemFields<T = never> = {
  id: RecordIdString;
  created: IsoDateString;
  updated: IsoDateString;
  collectionId: string;
  collectionName: Collections;
  expand?: T;
};
// ...
export type SuperiorRecord = {
  idTipoInstitucion?: RecordIdString;
  idUsuario?: RecordIdString;
  municipio?: string;
  nombre?: string;
  region?: string;
};

export type TipoBachillerRecord = {
  descripcion?: string;
};

export type TipoInstitucionRecord = {
  descripcion?: string;
};
// ...
export type SuperiorResponse<Texpand = unknown> = Required<SuperiorRecord> &
  BaseSystemFields<Texpand>;
export type TipoBachillerResponse<Texpand = unknown> =
  Required<TipoBachillerRecord> & BaseSystemFields<Texpand>;
export type TipoInstitucionResponse<Texpand = unknown> =
  Required<TipoInstitucionRecord> & BaseSystemFields<Texpand>;